Implement loading of jni native library from inside the jar.
The vw_jni jar already contains the native library for macos, but this
native library is not actually used. This change includes the logic to
extract this library from the jar and load it. This would be specially
useful if in the future the native libraries for linux and windows
were included as well inside the jar.